fix: fixed general fast bugs

+++ src/main/java/com/letraaletra/api/features/transaction/infrastructure/presentation/dto/response/transaction/TransactionResponse.java
@@ -13,9 +13,9 @@ public record TransactionResponse(
         String username,
 
         CoinType coinType,
-        int amount,
-        int balanceBefore,
-        int balanceAfter,
+        long amount,
+        long balanceBefore,
+        long balanceAfter,
 
         OperationType operation,
