+++ src/main/java/com/letraaletra/api/features/user/application/usecase/GrantUserRewardUseCase.java
@@ -81,9 +81,9 @@ public class GrantUserRewardUseCase implements UseCase<GrantUserRewardInput, Voi
                     user.getUserId(),
                     walletMovement.coinType(),
                     walletMovement.amount(),
-                    (int) walletMovement.balanceBefore()
+                    walletMovement.balanceBefore()
                             .getAmountFor(walletMovement.coinType()),
-                    (int) walletMovement.balanceAfter()
+                    walletMovement.balanceAfter()
                             .getAmountFor(walletMovement.coinType()),
                     walletMovement.operation(),
                     TransactionReason.ADMIN_GIVE,
