+++ src/main/java/com/letraaletra/api/features/user/infrastructure/service/UpdateStatsService.java
@@ -68,9 +68,9 @@ public class UpdateStatsService implements UserStatsService {
                             user.getUserId(),
                             walletMovement.coinType(),
                             walletMovement.amount(),
-                            (int) walletMovement.balanceBefore()
+                            walletMovement.balanceBefore()
                                     .getAmountFor(walletMovement.coinType()),
-                            (int) walletMovement.balanceAfter()
+                            walletMovement.balanceAfter()
                                     .getAmountFor(walletMovement.coinType()),
                             walletMovement.operation(),
                             TransactionReason.LEVEL_UP,
